adds interfaces back for design purposes
